feat: tratamento de excecoes C/D entidade User

+++ src/domains/User/services/UserService.ts
@@ -1,6 +1,8 @@
 import { User } from "@prisma/client";
 import  prisma  from "../../../../config/prismaClient";
 import bcrypt from "bcrypt";
+import { InvalidParamError } from "../../../../errors/InvalidParamError";
+import { QueryError } from "../../../../errors/QueryError";    
 
 class UserService {
     async encryptPassword(password: string) {
@@ -10,6 +12,20 @@ class UserService {
     }
 
     async create(body: User) {
+        const checkUser = await prisma.user.findUnique({
+            where: {
+                email: body.email
+            }
+        });
+      
+        if (!checkUser) {
+            throw new QueryError("Esse email já está cadastrado");
+        }
+
+        if (body.email == null) {
+            throw new InvalidParamError("Email não informado!");
+        }
+      
         const encrypted = await this.encryptPassword(body.password);
         const user = await prisma.user.create({
             data: {
@@ -20,11 +36,12 @@ class UserService {
                 role: body.role
             }
         });
+
         return user;
     }
 
     async readId(id: number){
-        const user = await prisma.user.findUnique ({
+        const user = await prisma.user.findUnique({
             where: {
                 id: id
             },
@@ -46,17 +63,35 @@ class UserService {
         return user;
     }
 
-    async update(body: User){
-        const user = await prisma.user.update ({
+    async update(body: User) {
+        const checkUser = await prisma.user.findUnique({
+            where: {
+                id: body.id
+            },
+        });
+        if (!checkUser) {
+            throw new QueryError("Usuário não cadastrado no sistema!");
+        }
+
+        const user = await prisma.user.update({
             where: {
                 id: body.id
             },
             data: body
         });
+
         return user;
     }
 
-    async deleteId(id:number){
+    async deleteId(id: number) {
+        const checkUser = prisma.user.findUnique({
+            where: {
+                id: id
+            }
+        });
+        if (!checkUser) {
+            throw new QueryError("Usuário não existe.");
+        }
         const user = await prisma.user.delete({
             where: {
                 id: id
